Update banner.html to announce the Q2 survey (#10731)
_Description of what this PR is changing or adding, and why:_ Update the
banner to include a link to the Q2 survey. To be removed on the 18th.

_Issues fixed by this PR (if any):_ NA

_PRs or commits this PR depends on (if any):_ NA

## Presubmit checklist

- [x] This PR is marked as draft with an explanation if not meant to
land until a future stable release.
- [x] This PR doesn’t contain automatically generated corrections
(Grammarly or similar).
- [x] This PR follows the [Google Developer Documentation Style
Guidelines](https://developers.google.com/style) — for example, it
doesn’t use _i.e._ or _e.g._, and it avoids _I_ and _we_ (first person).
- [x] This PR uses [semantic line
breaks](https://github.com/dart-lang/site-shared/blob/main/doc/writing-for-dart-and-flutter-websites.md#semantic-line-breaks)
of 80 characters or fewer.