Fix `getOS` for all operating systems (#10737)
In the current website getOS() returns "macos" if the user agent is on
macOS or iOS and otherwise windows, since -1 is truthy, `if
(userAgent.indexOf('Win'))` is always true and so "windows" is always
returned. This PR fixes that behaviour as well as fixing an android
check that would only return linux if the device is android, rather than
the opposite.