add breadcrumb about flavor-conditional asset bundling to assets page (#10823)
This is a follow-up to
https://github.com/flutter/website/pull/9864#issuecomment-1831179497.

This adds a breadcrumb to the "Assets and images" page that mentions and
links to the flavor-conditional asset bundling feature. This is meant to
aid discovery of the feature for users who are new to the assets
feature.

I chose adding a simple bread crumb over a full explanation + example
pubspec as this is a relatively niche feature, and I didn't want there
to be a significant amount of text between the top-level "Specifying
assets" and "Loading assets" sections, as the beginning of these
sections are going to be what I feel like the vast-majority of visitors
are going to be interested in.

## Presubmit checklist

- [x] This PR is marked as draft with an explanation if not meant to
land until a future stable release.
- [x] This PR doesn’t contain automatically generated corrections
(Grammarly or similar).
- [x] This PR follows the [Google Developer Documentation Style
Guidelines](https://developers.google.com/style) — for example, it
doesn’t use _i.e._ or _e.g._, and it avoids _I_ and _we_ (first person).
- [x] This PR uses [semantic line
breaks](https://github.com/dart-lang/site-shared/blob/main/doc/writing-for-dart-and-flutter-websites.md#semantic-line-breaks)
of 80 characters or fewer.