Add Elementary to the state management options page (#11035)
This PR adds a section about the Elementary library to the state
management options page.
Recently MVVM was
[mentioned](https://docs.flutter.dev/get-started/fwe/state-management#using-mvvm-for-your-applications-architecture)
in the documentation as one of the ways to build Flutter apps.

Elementary is an implementation of the Model-View-ViewModel pattern for
Flutter applications. The library has been used by many teams for quite
some time already. The library is highly tested (100% code coverage and
test cases), and has detailed documentation as in code, as a separate
website.

It might be helpful for the community to mention the library in this
list.

## Presubmit checklist

- [x] This PR is marked as draft with an explanation if not meant to
land until a future stable release.
- [x] This PR doesn’t contain automatically generated corrections
(Grammarly or similar).
- [x] This PR follows the [Google Developer Documentation Style
Guidelines](https://developers.google.com/style) — for example, it
doesn’t use _i.e._ or _e.g._, and it avoids _I_ and _we_ (first person).
- [x] This PR uses [semantic line
breaks](https://github.com/dart-lang/site-shared/blob/main/doc/writing-for-dart-and-flutter-websites.md#semantic-line-breaks)
of 80 characters or fewer.

---------

Co-authored-by: Brett Morgan <brettmorgan@google.com>