Make it clear that relative base hrefs are supported for the client/runtime if appropriately formed
There are some places where a relative `base href` is not supported (and these are documented/validated), but there is nothing explicitly calling out that they are supported in the client app. This adds a note to solidify this so they can be used (for example in the DevTools server) without the worry of them being undocumented.

Fixes https://github.com/flutter/flutter/issues/154620