Add temporary Dash image to FWE index  (#11218)
This PR adds an image of Dash to the intro page of FWE, and updates the
title to match the image. It's important to add an image because social
posts about this page are going out on Friday. I consider this
temporary, and not a fix for the [existing
issue](https://github.com/flutter/website/issues/11194). That will
require more time.

I edited the image to have a transparent background, including the
inside of the magnifying glass, so it should work in light and dark
mode.

<img width="842" alt="Screenshot 2024-10-02 at 8 45 06 AM"
src="https://github.com/user-attachments/assets/b8fa179d-1809-4e7f-99b4-ef99d616ce37">