Add initial pages for Architecture guidance (#11300)
Adds the first 3 pages of the architecture guidance documentation. 

This PR **does not** add these pages to the side nav. There will be
another PR that fast follows this one and contains a couple more pages
of docs, as well as the the side nav.

FYI: These pages have been reviewed by @sfshaza2 in a Google Doc. Of
course additional reviews are welcomed.

URL for added pages:
https://flutter-docs-prod--pr11300-ew-app-architecture-first-xhdhrb5w.web.app/app-architecture

---------

Co-authored-by: Shams Zakhour (ignore Sfshaza) <44418985+sfshaza2@users.noreply.github.com>