Show correct video for LinearGradient widget. (#11420)
_Description of what this PR is changing or adding, and why:_
The same video was appearing twice. This PR updates the one of the
videos to display the correct one for the LinearGradient widget.

_Issues fixed by this PR (if any):_
[11288](https://github.com/flutter/website/issues/11288)

_PRs or commits this PR depends on (if any):_
n/a

## Presubmit checklist

- [ x] This PR is marked as draft with an explanation if not meant to
land until a future stable release.
- [ x] This PR doesn’t contain automatically generated corrections
(Grammarly or similar).
- [ x] This PR follows the [Google Developer Documentation Style
Guidelines](https://developers.google.com/style) — for example, it
doesn’t use _i.e._ or _e.g._, and it avoids _I_ and _we_ (first person).
- [ x] This PR uses [semantic line
breaks](https://github.com/dart-lang/site-shared/blob/main/doc/writing-for-dart-and-flutter-websites.md#semantic-line-breaks)
of 80 characters or fewer.