Add a runtime switch guide to control Impeller usage
Impeller has been causing issues on lower-end or older devices.
While stabilizing Impeller remains a priority, it is essential
to provide app developers with runtime control over its usage.
This allows decisions to be made on a per-device or per-SoC basis.

This patch introduces a guide for controlling Impeller usage at runtime
using the `getFlutterShellArgs()` method.