Remove flutter_gdb from docs (it's deprecated). (#11679)
_Description of what this PR is changing or adding, and why:_

I recently discovered that GDB for Android has been deprecated. Triaged
the issue with the Flutter dev team and the resolution was to remove
mention of the flutter_gdb tool from the docs. There is something newer
called LLDB, but we don't have a tool for that.

_Issues fixed by this PR (if any):_

_PRs or commits this PR depends on (if any):_

## Presubmit checklist

- [x] This PR is marked as draft with an explanation if not meant to
land until a future stable release.
- [x] This PR doesn’t contain automatically generated corrections
(Grammarly or similar).
- [x] This PR follows the [Google Developer Documentation Style
Guidelines](https://developers.google.com/style) — for example, it
doesn’t use _i.e._ or _e.g._, and it avoids _I_ and _we_ (first person).
- [x] This PR uses [semantic line
breaks](https://github.com/dart-lang/site-shared/blob/main/doc/writing-for-dart-and-flutter-websites.md#semantic-line-breaks)
of 80 characters or fewer.