Update integration testing docs for Android and iOS. (#11680)
_Description of what this PR is changing or adding, and why:_

Separate the iOS and Android integration testing steps so that it's
easier for our Android and iOS devs to parse this content.

Notes:

I noticed some issues with the guide while making the original updates
and made a few updates to fix the issues. This includes:
- Strengthened the introduction.
- Removed the 'recipe' wording (this isn't structured like our recipes
nor in the recipe directory).
- There were quite a few coding errors and grammatical flow problems in
the original Firebase Test Lab section for Android. I've cleaned these
up but I'm also not 100% certain that those commands are accurate today.
We might want to follow up on that section and make sure the steps work.

_Issues fixed by this PR (if any):_

_PRs or commits this PR depends on (if any):_

## Presubmit checklist

- [x] This PR is marked as draft with an explanation if not meant to
land until a future stable release.
- [x] This PR doesn’t contain automatically generated corrections
(Grammarly or similar).
- [x] This PR follows the [Google Developer Documentation Style
Guidelines](https://developers.google.com/style) — for example, it
doesn’t use _i.e._ or _e.g._, and it avoids _I_ and _we_ (first person).
- [x] This PR uses [semantic line
breaks](https://github.com/dart-lang/site-shared/blob/main/doc/writing-for-dart-and-flutter-websites.md#semantic-line-breaks)
of 80 characters or fewer.