Update a few deep link notes to support iOS (#11687)
_Description of what this PR is changing or adding, and why:_

We had a few places in the documentation that still suggestion deep link
validation was only supported by Android. Updated these to include iOS.

Notes:
Updated the section called How to launch DevTools so that it's easier to
parse, but feel free to revert that change if meaning is lost.

_Issues fixed by this PR (if any):_

_PRs or commits this PR depends on (if any):_

## Presubmit checklist

- [x] This PR is marked as draft with an explanation if not meant to
land until a future stable release.
- [x] This PR doesn’t contain automatically generated corrections
(Grammarly or similar).
- [x] This PR follows the [Google Developer Documentation Style
Guidelines](https://developers.google.com/style) — for example, it
doesn’t use _i.e._ or _e.g._, and it avoids _I_ and _we_ (first person).
- [x] This PR uses [semantic line
breaks](https://github.com/dart-lang/site-shared/blob/main/doc/writing-for-dart-and-flutter-websites.md#semantic-line-breaks)
of 80 characters or fewer.