Update "Obfuscate your app" topic to be more platform agnostic. (#11690)
_Description of what this PR is changing or adding, and why:_

This topic only had Android code snippets. To make it easier for devs
creating apps for other platforms, I reformatted the commands to be
platform agnostic.

Notes:
- We had a few limitations outside of the Limitations section, so moved
them there.
- Noticed we had a warning inside of the Limitations section, so updated
the heading to include warnings.
- Updated some of the variable names in the commands and created a
bulleted list below each command with the variable names and what they
need to be replaced with.
- 'Read an obfuscated name' contents looked like they needed a revision,
so took care of that. Essentially it looks like that section is about
generating an obfuscation map that contains keys and values. It's still
not quite right, but I hope it's closer to where things are at now. I'm
also not sure if we still need the `--extra-gen-snapshot-options=`
before the `--save-obfuscation-map=` command.

_Issues fixed by this PR (if any):_

_PRs or commits this PR depends on (if any):_

## Presubmit checklist

- [x] This PR is marked as draft with an explanation if not meant to
land until a future stable release.
- [x] This PR doesn’t contain automatically generated corrections
(Grammarly or similar).
- [x] This PR follows the [Google Developer Documentation Style
Guidelines](https://developers.google.com/style) — for example, it
doesn’t use _i.e._ or _e.g._, and it avoids _I_ and _we_ (first person).
- [x] This PR uses [semantic line
breaks](https://github.com/dart-lang/site-shared/blob/main/doc/writing-for-dart-and-flutter-websites.md#semantic-line-breaks)
of 80 characters or fewer.