Update performance profiling docs. (#11688)
_Description of what this PR is changing or adding, and why:_

This is part of one of our goals to make it easier for iOS developers to
work with our docs. iOS devs are having a particularly difficult time
with debugging and performance. This guide is full of great information,
but it's been added to over the years by many contributors and looks
like it has started to lose its original form.

I've taken a shot at condensing all of the materials that have been
added over the years and making this doc easier to parse in a linear
fashion. I worked on it a bit more than intended, so if the changes are
too much in one PR, no worries at all. I can always split this up.

Changes:

- Created an overview and moved the various notes and conceptual details
into this overview.
- The overview now has a list of performance analysis tools that the
developer can browse with links to the documentation elsewhere.
- Tried to make it more clear that this page is specifically about the
Performance Overlay tool, but more might be needed.
- A few sections in the doc weren't about the Performance Overlay tool,
so I moved them into the bulleted list in the overview with links to
their documentation (the sections were very small, so this was pretty
easy).
- Changed some headings, and moved some sections around.
- Moved all of the links to the bottom of the topic and alphabetically
sorted them (to match the pattern i've seen elsewhere the most often.)

_Issues fixed by this PR (if any):_

_PRs or commits this PR depends on (if any):_

## Presubmit checklist

- [x] This PR is marked as draft with an explanation if not meant to
land until a future stable release.
- [x] This PR doesn’t contain automatically generated corrections
(Grammarly or similar).
- [x] This PR follows the [Google Developer Documentation Style
Guidelines](https://developers.google.com/style) — for example, it
doesn’t use _i.e._ or _e.g._, and it avoids _I_ and _we_ (first person).
- [x] This PR uses [semantic line
breaks](https://github.com/dart-lang/site-shared/blob/main/doc/writing-for-dart-and-flutter-websites.md#semantic-line-breaks)
of 80 characters or fewer.