Update iOS and Android docs for flavors. (#11709)
_Description of what this PR is changing or adding, and why:_

- We now have flavors.md (Android) and flavors-ios.md (iOS and macOS)
- Moved iOS workflow from flavors.md to flavors-ios.md
- Both documents now focus on the build settings 'production' and
'staging' flavors instead of 'free' and 'premium'.
- I have made minimal changes to the Android workflow in this PR and
have focused on just making the general wording we use the same for
Android and iOS. After this CL is submitted, I'll create a PR to update
the Android workflow.
- @Sfshaza I tried something new with code samples. I added a title
above each of them to make it clear where a change needs to be made
(console, podfile, etc). It helped ground me, but it might be overkill.
Go ahead and remove titles if you feel they aren't needed.
- @vashworth, can you check my iOS and macOS changes?
- @reidbaker, can you check my terminology changes for Android. Ignore
everything else. I'll have more updates in the next PR, which will be
specifically for Android.

_Issues fixed by this PR (if any):_

I'll add the list to the next PR. There are a lot of them.

_PRs or commits this PR depends on (if any):_

## Presubmit checklist

- [x] This PR is marked as draft with an explanation if not meant to
land until a future stable release.
- [x] This PR doesn’t contain automatically generated corrections
(Grammarly or similar).
- [x] This PR follows the [Google Developer Documentation Style
Guidelines](https://developers.google.com/style) — for example, it
doesn’t use _i.e._ or _e.g._, and it avoids _I_ and _we_ (first person).
- [x] This PR uses [semantic line
breaks](https://github.com/dart-lang/site-shared/blob/main/doc/writing-for-dart-and-flutter-websites.md#semantic-line-breaks)
of 80 characters or fewer.