Add a breaking change notice for `integration_test` golden-file comparisons (#11716)
While this is more of a bug fix than a breaking change, it _could_ break
users that were expecting it not to work and have implemented
workarounds. At the very least it gives a canonical place to point to
"this is fixed and here is what is wrong" for customers to use.

---------

Co-authored-by: Shams Zakhour (ignore Sfshaza) <44418985+sfshaza2@users.noreply.github.com>