Show set up entries in sidenav when child page is open (#11732)
An improvement for the flow of someone navigating between the set up
pages in the sidenav. Before this PR, they didn't show up at all in the
sidenav, but now they do if any of the set up pages are open.

Contributes to https://github.com/flutter/website/issues/11672 as an
improvement while I work on a longer term, automated solution.