Introduce new "On this page" design (#11701)
- Update the design for the right-side TOC on wide screens, that uses a
blue visual to indicate the currently visible headers.
- Introduce a completely new design for the narrow screen TOC that
persists at the top of the page and opens a popup when clicked to
navigate to other headers on the page. This is inspired by various other
documentation sites that do something similar now and will be familiar
to developers reading the docs.

**Testing:** Try interacting with the TOC in both wide and narrow
layouts on a variety of pages. Consider trying the [layout overview
page](https://flutter-docs-prod--pr11701-feat-on-this-page-hukxosz9.web.app/ui/layout)
which has many headers.

Resolves https://github.com/flutter/website/issues/11750