Update flavors guide for Android. (#11758)
_Description of what this PR is changing or adding, and why:_

This is a refactor of the Android Flavors guide that addresses several
developer requests, including steps to customize the app name and app
icons.

_Issues fixed by this PR (if any):_

Fixes: #11591 #11060 #10713 #10309 #10164 #9976 #9360 #9359 #9309 #8743
#8746 #8238

_PRs or commits this PR depends on (if any):_

This is part of a larger cleanup effort for the flavors documentation.
The first part of this cleanup addressed our iOS and macOS docs for
flavors. This was the final part of the cleanup.

## Presubmit checklist

- [x] This PR is marked as draft with an explanation if not meant to
land until a future stable release.
- [x] This PR doesn’t contain automatically generated corrections
(Grammarly or similar).
- [x] This PR follows the [Google Developer Documentation Style
Guidelines](https://developers.google.com/style) — for example, it
doesn’t use _i.e._ or _e.g._, and it avoids _I_ and _we_ (first person).
- [x] This PR uses [semantic line
breaks](https://github.com/dart-lang/site-shared/blob/main/doc/writing-for-dart-and-flutter-websites.md#semantic-line-breaks)
of 80 characters or fewer.