Reference alternative Linux deploy methods (#11840)
_Description of what this PR is changing or adding, and why:_

This PR follows up on a discussion about alternative Linux deploy
methods, and a comment that was made about it:
https://github.com/flutter/flutter/issues/160430#issuecomment-2597270972

This is specifically about the second bullet point:

> If documentation exists elsewhere (e.g. upstream to that packaging
format or a good blog post) then we should link to that from the Flutter
documentation. That helps developers use these packaging formats without
overcomplicating the Flutter documentation.

To be clear about my role in this play, I am the developer of the
flatpak-flutter tool.

_Issues fixed by this PR (if any):_
https://github.com/flutter/website/issues/7639
https://github.com/flutter/website/issues/7641

_PRs or commits this PR depends on (if any):_

## Presubmit checklist

- [X] This PR is marked as draft with an explanation if not meant to
land until a future stable release.
- [X] This PR doesn’t contain automatically generated corrections
(Grammarly or similar).
- [X] This PR follows the [Google Developer Documentation Style
Guidelines](https://developers.google.com/style) — for example, it
doesn’t use _i.e._ or _e.g._, and it avoids _I_ and _we_ (first person).
- [X] This PR uses [semantic line
breaks](https://github.com/dart-lang/site-shared/blob/main/doc/writing-for-dart-and-flutter-websites.md#semantic-line-breaks)
of 80 characters or fewer.