Remove outdated trailing comma mention (#11886)
This section is no longer relevant as the formatter automatically adds
or removes trailing commas as necessary in Dart 3.7 and later.

The images also include inaccessible text, so just remove it and point
to the kept up to date dart.dev docs.

Contributes to https://github.com/flutter/website/issues/11883