Add samples index page  (#11815)
_Description of what this PR is changing or adding, and why:_

Adds a new page that lists all samples, codelabs, and cookbook recipes.
Samples are under-utilized, and this makes them more discoverable.
There's also a plan to retire the existing stand-alone Samples Index.

Each resource will have an associated GA custom event, which will give
us a way to compare usage of samples, codelabs and recipes (sort of).

For now, the Cookbook and Codelabs page aren't removed. and therefor
content is duplicated on this page. I don't intend on suggesting that we
remove those pages unless it's clear that folks are using this new page
to get to codelabs.

Still to-do:
- [x] A bunch of data entry. 
- [x] Add the GA custom event logic.

_Issues fixed by this PR (if any):_
* #9185 
* [samples 2582 ](https://github.com/flutter/samples/issues/2582)
* [samples 2507](https://github.com/flutter/samples/issues/2507)


## Presubmit checklist

- [x] This PR is marked as draft with an explanation if not meant to
land until a future stable release.
- [x] This PR doesn’t contain automatically generated corrections
(Grammarly or similar).
- [x] This PR follows the [Google Developer Documentation Style
Guidelines](https://developers.google.com/style) — for example, it
doesn’t use _i.e._ or _e.g._, and it avoids _I_ and _we_ (first person).
- [x] This PR uses [semantic line
breaks](https://github.com/dart-lang/site-shared/blob/main/doc/writing-for-dart-and-flutter-websites.md#semantic-line-breaks)
of 80 characters or fewer.

---------

Co-authored-by: Parker Lougheed <parlough@gmail.com>