Add standalone 'Install using VS Code' page (#11992)
For now, it's not linked to anywhere or in the sidenav, but that will
change once some other necessary content lands.

Part of https://github.com/flutter/website/issues/11911

Some of this content is duplicated from elsewhere. Once every part of
this issue lands, some pieces might be extracted to partials where
helpful.

**Staged:**
https://flutter-docs-prod--pr11992-feat-install-with-vs-code-6qmcg59p.web.app/install/with-vs-code