Add missing instructions for Android a2a depending on a module as source case (#12050)
_Description of what this PR is changing or adding, and why:_

The instruction to add 
```kotlin
maven("$storageUrl/download.flutter.io")
```
to your `dependencyResolutionManagement.repositories` section in the
settings gradle file was already included in the tab for setting up
depending on a module as an aar. It is also required for depending on
the module as source, so I pulled it up to the preceding section (right
before those tabs). I also made that section have valid Groovy and
Kotlin, instead of just providing Groovy.

_Issues fixed by this PR (if any):_

https://github.com/flutter/flutter/issues/158583

_PRs or commits this PR depends on (if any):_

N/A

## Presubmit checklist

- [ ] This PR is marked as draft with an explanation if not meant to
land until a future stable release.
- [ ] This PR doesn’t contain automatically generated corrections
(Grammarly or similar).
- [ ] This PR follows the [Google Developer Documentation Style
Guidelines](https://developers.google.com/style) — for example, it
doesn’t use _i.e._ or _e.g._, and it avoids _I_ and _we_ (first person).
- [ ] This PR uses [semantic line
breaks](https://github.com/dart-lang/site-shared/blob/main/doc/writing-for-dart-and-flutter-websites.md#semantic-line-breaks)
of 80 characters or fewer.

Co-authored-by: Gray Mackall <mackall@google.com>