Add explanation for edge-to-edge Google Play Console warning (#12118)
Adds explanation about warnings that users may see in the Google Play
Console related to edge-to-edge mode. Added the explanation to
[this](https://github.com/flutter/website/blob/449f33220f4d95d05c4f942be6e21e1036cadf38/src/content/release/breaking-changes/default-systemuimode-edge-to-edge.md)
page because Android added the warning to the console as part of
edge-to-edge mode becoming the default on Android 15+.

This addition was prompted by
https://github.com/flutter/flutter/issues/169810, which a user reported
because they were seeing such warning.

## Presubmit checklist

- [x] This PR is marked as draft with an explanation if not meant to
land until a future stable release.
- [x] This PR doesn’t contain automatically generated corrections
(Grammarly or similar).
- [ ] This PR follows the [Google Developer Documentation Style
Guidelines](https://developers.google.com/style) — for example, it
doesn’t use _i.e._ or _e.g._, and it avoids _I_ and _we_ (first person).
- [x] This PR uses [semantic line
breaks](https://github.com/dart-lang/site-shared/blob/main/doc/writing-for-dart-and-flutter-websites.md#semantic-line-breaks)
of 80 characters or fewer.