Remove references to Hive and Isar (#12196)
_Description of what this PR is changing or adding, and why:_

Hive and Isar are both dead packages. It is unadvisable to use them in
new projects. For now, I have replaced Hive with Hive CE, but since I am
the maintainer of Hive CE, I will leave it up to other maintainers to
determine if we should remove the mention of Hive entirely instead.

_Issues fixed by this PR (if any):_

N/A

_PRs or commits this PR depends on (if any):_

N/A

## Presubmit checklist

- [x] **If you are unwilling, or unable, to sign the CLA, even for a
_tiny_, one-word PR, please file an issue instead of a PR.**
- [x] If this PR is not meant to land until a future stable release,
mark it as draft with an explanation.
- [x] This PR doesn't contain automatically generated corrections
(generated by Grammarly or similar).
- [x] This PR follows the [Google Developer Documentation Style
Guidelines](https://developers.google.com/style)—for example, it doesn't
use _i.e._ or _e.g._, and it avoids _I_ and _we_ (first-person
pronouns).
- [x] This PR uses [semantic line
breaks](https://github.com/dart-lang/site-shared/blob/main/doc/writing-for-dart-and-flutter-websites.md#semantic-line-breaks)
  of 80 characters or fewer.