Remove "Component theme normalization updates" from 3.32 changes (#12266)
_Description of what this PR is changing or adding, and why:_ This PR
removes a duplicate link from the "Breaking changes" page.

_Issues fixed by this PR (if any):_ The link to "Component theme
normalization updates" appears in both "Not yet released to stable" and
"Released in Flutter 3.32" sections; however, since the change has not
been released in 3.32, it should not be listed in the 3.32 section.

_PRs or commits this PR depends on (if any):_ N/A

## Presubmit checklist

- [x] If you are unwilling, or unable, to sign the CLA, even for a
_tiny_, one-word PR, please file an issue instead of a PR.
- [x] If this PR is not meant to land until a future stable release,
mark it as draft with an explanation.
- [x] This PR follows the [Google Developer Documentation Style
Guidelines](https://developers.google.com/style)—for example, it doesn't
use _i.e._ or _e.g._, and it avoids _I_ and _we_ (first-person
pronouns).
- [x] This PR uses [semantic line
breaks](https://github.com/dart-lang/site-shared/blob/main/doc/writing-for-dart-and-flutter-websites.md#semantic-line-breaks)
  of 80 characters or fewer.