Fix: Add missing `windowOptOutEdgeToEdgeEnforcement` to all relevant styles for SDK 35 compatibility (#12280)
The Flutter documentation currently states that the
`<item name="android:windowOptOutEdgeToEdgeEnforcement">true</item>`
attribute should be added only to the `NormalTheme` in
`your_app/android/app/src/main/res/values/styles.xml`.

However, this is not sufficient.

To ensure proper behavior—especially for users with dark mode
enabled—this attribute must also be added to the `LaunchTheme` and be
mirrored in the `values-night/styles.xml` file.

Initially, I added the attribute only to `NormalTheme` in the `values`
folder, but users in dark mode continued to experience issues. This PR
resolves the problem by applying the fix consistently across both
`NormalTheme` and `LaunchTheme`, in both `values` and `values-night`.

---------

Co-authored-by: Parker Lougheed <parlough@gmail.com>