[3.35] Add a quick version of getting started (#12052)
**Dependent on hot reload for web not being behind a flag.**

Introduces a new get started flow that is shorter and focuses on
installing Flutter, VS Code, and Chrome then trying out Flutter and hot
reload on the web. New adding a new target platform docs have been added
in preparation for this and are linked to at the end of the doc.

Contributes to https://github.com/flutter/website/issues/11911

**Staged:**
https://flutter-docs-prod--pr12052-feat-get-started-quickly-zw2vhwsv.web.app/get-started/quick

---------

Co-authored-by: Amanda Fitch <18406675+antfitch@users.noreply.github.com>