Fix Android Studio installation link (#12359)
_Description of what this PR is changing or adding, and why:_ 

Fix the installation link to Android Studio (based on Google-internal UX
friction log).

_Issues fixed by this PR (if any):_

Currently we link to the Android Studio systems requirements from our
installation page. As a result, when clicking the Android Studio link on
our site you end up deep on a page and then have to scroll all the way
to the top to find the actual install step. This links directly to where
the installation link is.

_PRs or commits this PR depends on (if any):_

## Presubmit checklist

- [x] If you are unwilling, or unable, to sign the CLA, even for a
_tiny_, one-word PR, please file an issue instead of a PR.
- [x] If this PR is not meant to land until a future stable release,
mark it as draft with an explanation.
- [x] This PR follows the [Google Developer Documentation Style
Guidelines](https://developers.google.com/style)—for example, it doesn't
use _i.e._ or _e.g._, and it avoids _I_ and _we_ (first-person
pronouns).
- [x] This PR uses [semantic line
breaks](https://github.com/dart-lang/site-shared/blob/main/doc/writing-for-dart-and-flutter-websites.md#semantic-line-breaks)
  of 80 characters or fewer.