Update app-extensions.md (#12308)
**Description:**

In some environments the value of the Minimum Deployments is not set for
the Runner Target, and the ShareExtension one takes the latest one it
can, so they not match causing the Extension not to be shown on the
Photos Share sheet.

So I added an extra step for people following the tutorial to check that
small detail that can cause frustration.

**Issues solved**: None
**Commits dependency**: None