Fix: Update URL launch method to use Uri.parse (#12518)
Fixes #12410 

**Before:**
- Wasn't opening the website by clicking on the launch button. I don't
have a screenshot of that as the app was not breaking but was just not
opening the webpage. I could not find any error logs in the emulator
logs. Only thing I can verify is that this issue persists on macOS and iPhone 14 emulator.

**After:**
Opened the website successfully
<img width="382" alt="image"
src="https://github.com/user-attachments/assets/455279e2-b1f5-4d5b-b50c-fef7b4521ab5"
/>