Remove package dependency in FWE part 1 (#12561)
_Description of what this PR is changing or adding, and why:_

FWE was written to depend on a package to hide the game logic from the
reader. I no longer thing that's a good idea. It'll make it tougher for
AI to understand the entire context, and it'll require me to maintain a
separate package.

This removes that dependency by adding an instruction to copy and paste
a big ol dart file at the beginning of FWE. I also adds a collapsible
panel to hide the code so it looks less ugly :)

## Presubmit checklist

- [x] If you are unwilling, or unable, to sign the CLA, even for a
_tiny_, one-word PR, please file an issue instead of a PR.
- [x] If this PR is not meant to land until a future stable release,
mark it as draft with an explanation.
- [x] This PR follows the [Google Developer Documentation Style
Guidelines](https://developers.google.com/style)—for example, it doesn't
use _i.e._ or _e.g._, and it avoids _I_ and _we_ (first-person
pronouns).
- [x] This PR uses [semantic line
breaks](https://github.com/dart-lang/site-shared/blob/main/doc/writing-for-dart-and-flutter-websites.md#semantic-line-breaks)
  of 80 characters or fewer.