Refactor Mockito example (#12558)
_Description of what this PR is changing or adding, and why:_
- Someone reported an issue in the example of Mockito that APIs of
`MockClient` has been updated and we need to pass a handler to it. After
I checked, I realized that it was a generated class when we use
annotations provided by Mockito. I added a note and renamed the
generated class name. I hope it removes the confusion.
- More details in the issue linked below.  

_Issues fixed by this PR (if any):_
 - Fixes #12116 

## Presubmit checklist

- [ ] If you are unwilling, or unable, to sign the CLA, even for a
_tiny_, one-word PR, please file an issue instead of a PR.
- [ ] If this PR is not meant to land until a future stable release,
mark it as draft with an explanation.
- [x] This PR follows the [Google Developer Documentation Style
Guidelines](https://developers.google.com/style)—for example, it doesn't
use _i.e._ or _e.g._, and it avoids _I_ and _we_ (first-person
pronouns).
- [x] This PR uses [semantic line
breaks](https://github.com/dart-lang/site-shared/blob/main/doc/writing-for-dart-and-flutter-websites.md#semantic-line-breaks)
  of 80 characters or fewer.