Fix broken links to changelog from release note pages (#12661)
The problem that is described in the linked issue seems to occur due to the
Templating system not working for the snippet `{{site.repo.flutter}}`.
What differentiates the release notes pages from others is the option
`skipTemplateRendering` in the front matter, which is set to `true`.
Fix this by replacing the usage of template variables with the underlying values.

Fixes https://github.com/flutter/website/issues/12660