Add folding support to code blocks (#12680)
This adds support for defining **folding ranges** in any code block.

Folding ranges can be defined by specifying `foldable` on the code block
and using `[*` and `*]` lines in the code, and will be transformed to
collapsible sections with a toggle marker and indicator line. Folding
ranges can be nested. By default folding ranges are open, but can be set
to initially closed using `[* -`.

When using line numbers, the folding range markers and lines will be
shifted to the right of the line numbers.