Refactor to more idiomatic Jaspr code (#12726)
This refactors some parts of the code to have more idiomatic Jaspr
components and reduce logic in `global_scripts.dart`.

- Implement a new `ComponentRef` system that in some capacity works
similar to server components (more limited and less elegant, but
functionally working). This is to unblock some improvements (some of
which are included in this PR) to the interactive components of the
website, as we can already start architecting towards server components
and write less manual dom logic.
- Change the `Dropdown` component to accept direct `toggle` and
`content` components, which simplifies it's implementation. Migrate
existing uses of `Dropdown`.
- Migrate the top TOC (renamed to `PageNav`) to use the `Dropdown`
component, as well as the new `ComponentRef`s. Split into separate
"site-subheader" and "pagenav" elements. Preparations for #12715
- Add a new `Tooltip` component and migrate the glossary tooltips to it
using the new `ComponentRef`s. Can be reused for #12600

---------

Co-authored-by: Parker Lougheed <parlough@gmail.com>