Add DownloadableSnippet component (#12733)
Adds a component for downloading a code snippet as a file. This is
displayed as a collapsed code block with additional action buttons in
the title.

Provides three options:
- Download directly using browser apis
- Copy code to clipboard
- Expand to show full code

<img width="939" height="260" alt="Bildschirmfoto 2025-11-21 um 15 24
22"
src="https://github.com/user-attachments/assets/c0ef5e2a-7a8a-4022-a513-baf5c19100d4"
/>

---

I also updated the `CopyButton` to extract its content from the actual
code element instead of taking it as an parameter, to avoid duplicating
the data in the output html.