Add forceSingleThreadedSkwasm option to Flutter web loader (#178406) (#12712)
Update initialization.md

This PR updates the Flutter web initialization documentation
(`initialization.md`) to include the `forceSingleThreadedSkwasm` ption
for `_flutter.loader.load()`. This option allows developers to force the
Skia WebAssembly (skwasm) renderer to run in single-threaded mode, which
improves compatibility in browsers that do not support multi-threaded
WASM or when required security headers are missing.

- Related to -
[#178406](https://github.com/flutter/flutter/issues/178406)


## Presubmit checklist

- [ ] If you are unwilling, or unable, to sign the CLA, even for a
_tiny_, one-word PR, please file an issue instead of a PR.
- [ ] If this PR is not meant to land until a future stable release,
mark it as draft with an explanation.
- [x] This PR follows the [Google Developer Documentation Style
Guidelines](https://developers.google.com/style)—for example, it doesn't
use _i.e._ or _e.g._, and it avoids _I_ and _we_ (first-person
pronouns).
- [x] This PR uses [semantic line
breaks](https://github.com/dart-lang/site-shared/blob/main/doc/writing-for-dart-and-flutter-websites.md#semantic-line-breaks)
of 80 characters or fewer.